[BE] refactor: AuthDto와 GithubUserDto 이름 통일
