+++ server/src/domains/auth/api/AuthContorller.ts
@@ -38,23 +38,21 @@ export class AuthController {
   ) {
     if (!githubId) return;
     const userData = await this.authService.getUserProfile(githubId);
-    return { ...userData, role: role } as UserDto;
+    return { ...userData, role } as UserDto;
   }
 
-  @Get('/token')
-  @OpenAPI({ summary: 'Github OAuth 토큰을 발급받는 API' })
-  @ResponseSchema(GithubUserDto, { description: '토큰 정상 발급' })
-  async getGithubAccessToken(
-    @Session() session: any,
-    @SessionParam('githubId') githubId: string,
-    @QueryParam('code') code: string,
-  ) {
+  @Post('/login/social')
+  @OpenAPI({ summary: 'Github OAuth 로그인 API' })
+  @ResponseSchema(UserDto)
+  async socialLogin(@Session() session: any, @QueryParam('code') code: string) {
     const accessToken = await this.authService.getGithubAccessToken(code);
     const githubUserData = await this.authService.getGithubUserData(accessToken);
     const userData = await this.authService.findOrInsertUser(githubUserData);
-    if (!githubId) {
-      session.githubId = githubUserData.githubId;
-      session.role = 'MENTEE';
+    if (!session.user) {
+      session.user = {
+        githubId: githubUserData.githubId,
+        role: 'MENTEE',
+      };
     }
 
     return userData;
